Clean up some core memory / allocator code (#3547)

+++ src/sysdolphin/baselib/objalloc.h
@@ -6,8 +6,8 @@
 #include <sysdolphin/baselib/debug.h>
 
 typedef struct _objheap {
-    u32 top;
-    u32 curr;
+    uintptr_t top;
+    uintptr_t curr;
     u32 size;
     u32 remain;
 } objheap;
@@ -69,6 +69,7 @@ static inline void HSD_ObjAllocDisableNumLimit(HSD_ObjAllocData* data)
     data->num_limit_flag = 0;
 }
 
+/// A NULL pool uses HSD_MemAlloc for backing storage.
 void HSD_ObjSetHeap(u32 size, void* ptr);
 s32 HSD_ObjAllocAddFree(HSD_ObjAllocData* data, u32 num);
 void* HSD_ObjAlloc(HSD_ObjAllocData* data);
